Update README.md
GeoCentric is a Flask-based web application that utilizes the Google Maps API to help groups find the most centrally located chain (like a favorite restaurant, gym, etc.) among their members. The application determines this central location by finding the place that minimizes the standard deviation of travel times from each individual to the chain location.

This is perfect for friends, colleagues, or teams who are spread out geographically but want to find a common meeting spot that is fair in terms of travel time.